# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Management API",
    "description": "Backend API services for Project Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "Dashboard Analytics"
        ],
        "summary": "Get Unified Dashboard Telemetry for current user",
        "operationId": "get_dashboard_telemetry_api_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Dashboard Telemetry Api Dashboard Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/dashboard": {
      "get": {
        "tags": [
          "Dashboard Analytics"
        ],
        "summary": "Get Unified Dashboard Telemetry (Alias)",
        "operationId": "get_dashboard_telemetry_alias_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Dashboard Telemetry Alias Dashboard Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/members/users": {
      "get": {
        "tags": [
          "Project Members"
        ],
        "summary": "Get all users for dropdown",
        "description": "Returns all users as value/label items for a multi-select dropdown.",
        "operationId": "get_users_dropdown_api_members_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserDropdownItem"
                  },
                  "type": "array",
                  "title": "Response Get Users Dropdown Api Members Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/members/{project_id}": {
      "get": {
        "tags": [
          "Project Members"
        ],
        "summary": "Get project members",
        "description": "Returns all active members for a project (Enforces PBAC Project Visibility).",
        "operationId": "get_project_members_api_members__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectMemberResponse"
                  },
                  "title": "Response Get Project Members Api Members  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/members": {
      "post": {
        "tags": [
          "Project Members"
        ],
        "summary": "Add members to a project",
        "description": "Adds multiple members to a project (Enforces PBAC Project Update - Owner only).",
        "operationId": "add_members_api_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddMembersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddMembersRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "user_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "User Ids"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "user_ids"
        ],
        "title": "AddMembersRequest",
        "description": "Payload to add multiple members to a project."
      },
      "AddMembersResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "added": {
            "type": "integer",
            "title": "Added"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/ProjectMemberResponse"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "message",
          "added",
          "skipped",
          "members"
        ],
        "title": "AddMembersResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectMemberResponse": {
        "properties": {
          "project_member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Member Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "joined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "project_member_id",
          "project_id",
          "user_id",
          "full_name",
          "email",
          "is_active"
        ],
        "title": "ProjectMemberResponse",
        "description": "A single project member record."
      },
...
      "UserDropdownItem": {
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid",
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "title": "UserDropdownItem",
        "description": "For multi-select dropdown: value/label pair."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}